+++ source/android/android_time.c
@@ -49,6 +49,6 @@ uint64_t Sys_Microseconds( void ) {
 /*
 * Sys_Milliseconds
 */
-unsigned int Sys_Milliseconds( void ) {
+int64_t Sys_Milliseconds( void ) {
 	return Sys_Microseconds() / ( ( uint64_t )1000 );
 }
