+++ source/qcommon/qcommon.h
@@ -146,11 +146,11 @@ void SNAP_ParseBaseline( msg_t *msg, entity_state_t *baselines );
 void SNAP_SkipFrame( msg_t *msg, struct snapshot_s *header );
 struct snapshot_s *SNAP_ParseFrame( msg_t *msg, struct snapshot_s *lastFrame, int *suppressCount, struct snapshot_s *backup, entity_state_t *baselines, int showNet );
 
-void SNAP_WriteFrameSnapToClient( struct ginfo_s *gi, struct client_s *client, msg_t *msg, unsigned int frameNum, unsigned int gameTime,
+void SNAP_WriteFrameSnapToClient( struct ginfo_s *gi, struct client_s *client, msg_t *msg, int64_t frameNum, int64_t gameTime,
 								  entity_state_t *baselines, struct client_entities_s *client_entities,
 								  int numcmds, gcommand_t *commands, const char *commandsData );
 
-void SNAP_BuildClientFrameSnap( struct cmodel_state_s *cms, struct ginfo_s *gi, unsigned int frameNum, unsigned int timeStamp,
+void SNAP_BuildClientFrameSnap( struct cmodel_state_s *cms, struct ginfo_s *gi, int64_t frameNum, int64_t timeStamp,
 								struct fatvis_s *fatvis, struct client_s *client,
 								game_state_t *gameState, struct client_entities_s *client_entities,
 								bool relay, struct mempool_s *mempool );
@@ -764,10 +764,10 @@ extern cvar_t *versioncvar;
 extern cvar_t *revisioncvar;
 
 // host_speeds times
-extern unsigned int time_before_game;
-extern unsigned int time_after_game;
-extern unsigned int time_before_ref;
-extern unsigned int time_after_ref;
+extern int64_t time_before_game;
+extern int64_t time_after_game;
+extern int64_t time_before_ref;
+extern int64_t time_after_ref;
 
 /*
 ==============================================================
@@ -868,7 +868,7 @@ void    Sys_InitDynvars( void );
 
 void    Sys_AppActivate( void );
 
-unsigned int    Sys_Milliseconds( void );
+int64_t    Sys_Milliseconds( void );
 uint64_t        Sys_Microseconds( void );
 void        Sys_Sleep( unsigned int millis );
 
@@ -928,7 +928,7 @@ void SCR_BeginLoadingPlaque( void );
 void SV_Init( void );
 void SV_Shutdown( const char *finalmsg );
 void SV_ShutdownGame( const char *finalmsg, bool reconnect );
-void SV_Frame( int realMsec, int gameMsec );
+void SV_Frame( unsigned realMsec, unsigned gameMsec );
 bool SV_SendMessageToClient( struct client_s *client, msg_t *msg );
 void SV_ParseClientMessage( struct client_s *client, msg_t *msg );
