+++ source/snd_openal/snd_al.c
@@ -25,7 +25,7 @@ ALCdevice *alDevice = NULL;
 ALCcontext *alContext = NULL;
 
 #define UPDATE_MSEC 10
-static unsigned s_last_update_time;
+static int64_t s_last_update_time;
 
 int s_attenuation_model = 0;
 float s_attenuation_maxdistance = 0;
@@ -699,7 +699,7 @@ static pipeCmdHandler_t sndCmdHandlers[SND_CMD_NUM_CMDS] =
 */
 static int S_EnqueuedCmdsWaiter( sndCmdPipe_t *queue, pipeCmdHandler_t *cmdHandlers, bool timeout ) {
 	int read = S_ReadEnqueuedCmds( queue, cmdHandlers );
-	unsigned now = trap_Milliseconds();
+	int64_t now = trap_Milliseconds();
 
 	if( read < 0 ) {
 		// shutdown
